Add: Service optimization
